Force add "." to $PATH
As far as we call pngout and other binaries without specifying path to them, we have to add "." to $PATH variable to be sure that everything will works even if user installs IMGO not to /usr/local/bin, e.g. to ~/bin/ (which is not in $PATH from-the-box on some OS).